Crate conrod_wgpu

Structs

GlyphCacheCommand
An command for uploading an individual glyph.

Image
A loaded wgpu texture and it's width/height

Render
A render produced by the Renderer::render method.

Renderer
A helper type aimed at simplifying the rendering of conrod primitives via wgpu.

Vertex
The data associated with a single vertex.

Enums

RenderPassCommand
A draw command that maps directly to the wgpu::CommandEncoder method. By returning RenderPassCommands, we can avoid consuming the entire AutoCommandBufferBuilder itself which might not always be available from APIs that wrap Vulkan.